requirement.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
blinker==1.6.2
Bottleneck==1.3.5
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
discord.py==2.3.2
distlib==0.3.7
et-xmlfile==1.1.0
filelock==3.12.3
Flask==2.3.3
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gspread==5.10.0
httplib2==0.22.0
idna==3.4
iso8601==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
meson==1.2.1
multidict==6.0.4
numexpr==2.8.4
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.3
passlib==1.7.4
pip==23.2.1
platformdirs==3.10.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodomex==3.18.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyseto==1.7.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
PyYAML==6.0.1
replit==3.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==68.0.0
six==1.16.0
soupsieve==2.4.1
table2ascii==1.1.2
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.4
wcwidth==0.2.6
Werkzeug==2.3.7
wheel==0.38.4
yarl==1.9.2
